(1)重新定义实体类orders: extend_id 数据类型(2)修改UserAgentServiceImpl.java中根据用户ID获取所对应代理商函数

+++ ecp/ecp-dao/src/main/java/com/ecp/entity/Orders.java
@@ -43,13 +43,15 @@ public class Orders {
     private Date deleteTime;
     
     @Column(name = "extend_id")
-    private BigDecimal extendId;
+    private Long extendId;
+    
+    
 
-    public BigDecimal getExtendId() {
+	public Long getExtendId() {
 		return extendId;
 	}
 
-	public void setExtendId(BigDecimal extendId) {
+	public void setExtendId(Long extendId) {
 		this.extendId = extendId;
 	}
